test: move log formatter to util
